Reduce expensive calls to ioloop.IOLoop.instance()
IOLoop.instance() is only meant to be used when posting back to the
main thread, invoking this method performs double checks on a critical
code section.

It is better to use IOLoop.current(), which could in turn use .instance()
only if it were necessary.

I've debugged our code and we only have 3 threads running (according to my PyCharm debugger)

Therefore in most of the cases we're probably making those calls from the
same thread where Tornado's ioLoop thread lives.

I've also reduced the number of calls to IOLoop.current() by keeping a reference
to the loop in the constructor of the classes that need to communicate with it.
In all classes this reference is known as `self.loop`