Peers should only relay reachable close nodes
This fix implements a new constant for close node info timeout. Right now if your node issues a key search against the DHT and it's not found by a peer they will return close nodes no matter if they're reachable or not. This imposes a 15 minute limit for when the peer was last seen. This should cut down on peers sending you stale lists of nodes to populate your sidebar.

Modify to 5 Minutes

15 minutes seems a little long so let's go with 5